MY-APP/Vote/requirements.txt
sflask
redis
